[#194] Refactor
