+++ app/workers/github/deploy_worker.rb
@@ -26,12 +26,8 @@ module Github
       bot = Genova::Slack::Interactive::Bot.new
       bot.detect_github_event(deploy_job: deploy_job, commit_url: values[:commit_url], author: values[:author])
 
-      transaction = Genova::Utils::DeployTransaction.new(values[:repository])
-      transaction.begin
-
       Genova::Run.call(deploy_job)
 
-      transaction.commit
       bot.finished_deploy(deploy_job: deploy_job)
     rescue => e
       slack_notify(e)
