+++ app/workers/slack/deploy_target_worker.rb
@@ -6,7 +6,7 @@ module Slack
       logger.info('Started Slack::DeployTargetWorker')
 
       params = Genova::Slack::SessionStore.load(id).params
-      transaction = Genova::Utils::DeployTransaction.new(params[:repository])
+      transaction = Genova::TransactionManager.new(params[:repository])
 
       bot = Genova::Slack::Interactive::Bot.new(parent_message_ts: id)
       bot.ask_target(params)
